overlay: core: config: Enable night display on all platforms.
This boolean is required to show and enable the night display settings.
Not all platforms strictly adhere to the power/performance requirements,
for example Loire requires a fallback to GPU transformations because the
reported HWC2.0 functionality doesn't work, and a similar issue arises
on Nile SDE that has yet to be solved. The power hit doesn't seem nearly
as bad, and enabling it allows more testing and eventually pinning the
issue in libsdm-color that prevents the hardware transformation from
happening.

Signed-off-by: MarijnS95 <marijns95@gmail.com>