uinput-fpc.kl: Swap NAVIGATION_{LEFT,RIGHT} to match input header.
According the the input header, directional keys are defined as
follows:

define KEY_UP 103
define KEY_LEFT 105
define KEY_RIGHT 106
define KEY_DOWN 108

Swap the numbers to match. Do note that there is only a semantic
difference here; the numbers are swapped in the HAL as well, and do not
really have a meaning  ied to the header in the first place (they can
be any random number given that they are correctly mapped in a key
layout file). Yet, consistency is key.

Change-Id: I1ebe98858d033755abc222787b7623b825fba794
Signed-off-by: MarijnS95 <marijns95@gmail.com>