hardware: camera: Convert vendor->odm symlink to ro.hardware property.
Android 10 adds a nice feature that verifies whether the realpath of a
loaded module starts with the searched path the module came from. For
symlinks this is not true (for example the realpath of the camera
library is /odm/lib/hw/camera.qcom.so: this does _not_ start with
/vendor/lib/hw where the search originated from).

See the relevant change in libhardware:
https://android.googlesource.com/platform/hardware/libhardware/+/e448a05fecab601c59b89a2f948e1ad18050e40f%5E%21/#F0

Instead of creating a symlink on the odm partition, set
ro.hardware.camera=qcom in the relevant platform repositories (currently
only Tama and Kumano). This ensures libhardware also looks for
camera.qcom.so in all search paths (vendor and odm libraries).

Signed-off-by: MarijnS95 <marijns95@gmail.com>