liblights: Reintroduce corrected brighness bit-shifting routine.
This time the routine is implemented in a generic way, supporting any
number of bits between 8 and 15. Notice that second shift is corrected
(as explained in the previous commit) to shift
brightness >> 8 - mBacklightShift
instead of
brightness >> mBacklightShift
which would cause jumps in the output and would not be able to reach max
brightness on panels with more than 12 bits precision.

Signed-off-by: MarijnS95 <marijns95@gmail.com>