liblights: Always write brightness before enabling breath
A new kernel "feature" scales the breathing pattern by a previously
written brightness value, to blink in the right color instead of always
white.

Besides, not enabling breath if a color channel is not enabled is a good
idea anyway: this would have "fixed" always white led before that kernel
patch (assuming the given color is a pure red, green or blue color).